requirements.txt
WTForms~=3.2.1
SQLAlchemy~=2.0.49
Flask-Login~=0.6.3
Werkzeug~=3.1.8
Flask~=3.1.3
python-dotenv~=1.2.2